+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes to vueform
 
+## 1.2.0
+
+* Added [#3: Rename $updateCustomValidity to $setCustomValidity](https://github.com/optick/vueform/issues/3)
+* Added [#4: Simplify $updateValidity by accepting a single result boolean or string](https://github.com/optick/vueform/issues/4)
+
 ## 1.1.0
 
 * Added [#1: Add polyfill for ValidityState.tooShort](https://github.com/optick/vueform/issues/1)